+++ tests/dataset_tests/MNIST/MnistTest.cpp
@@ -29,10 +29,10 @@ TEST_F(MnistTest, loadData)
 {
     ASSERT_EQ(data->sizeOfData, 784);
     ASSERT_EQ(data->numberOfLabel, 10);
-    ASSERT_EQ(data->sets[training].inputs.size(), 60000);
-    ASSERT_EQ(data->sets[training].labels.size(), 60000);
-    ASSERT_EQ(data->sets[snn::testing].inputs.size(), 10000);
-    ASSERT_EQ(data->sets[snn::testing].labels.size(), 10000);
+    ASSERT_EQ((int)data->sets[training].inputs.size(), 60000);
+    ASSERT_EQ((int)data->sets[training].labels.size(), 60000);
+    ASSERT_EQ((int)data->sets[snn::testing].inputs.size(), 10000);
+    ASSERT_EQ((int)data->sets[snn::testing].labels.size(), 10000);
     ASSERT_EQ(data->sets[snn::testing].numberOfTemporalSequence, 0);
     ASSERT_EQ(data->sets[snn::testing].numberOfTemporalSequence, 0);
     ASSERT_EQ(data->isValid(), 0);
@@ -125,8 +125,8 @@ TEST_F(MnistTest, multipleLayersNeuralNetwork)
         Convolution(1, 4, activation::sigmoid),
         FullyConnected(10)
     });
-    neuralNetwork.optimizer.learningRate = 0.03;
-    neuralNetwork.optimizer.momentum = 0.90;
+    neuralNetwork.optimizer.learningRate = 0.03f;
+    neuralNetwork.optimizer.momentum = 0.90f;
     neuralNetwork.startTraining(*data);
     neuralNetwork.waitFor(2_ep || 60_s);
     neuralNetwork.stopTraining();
